requirements.txt
appnope==0.1.0
arrow==0.12.0
arxiv==0.2.2
beautifulsoup4==4.6.0
bibtexparser==0.6.2
bs4==0.0.1
certifi==2017.11.5
chardet==3.0.4
click==6.7
crossrefapi==1.2.0
cssselect==1.0.1
decorator==4.1.2
doi2bib==0.3.0
eutils==0.3.2
feedparser==5.2.1
Flask==0.12.2
Flask-SQLAlchemy==1.0
Flask-WTF==0.9.4
future==0.16.0
gcloud==0.18.3
gitdata==0.0.0
googleapis-common-protos==1.5.3
gunicorn==18.0
httplib2==0.10.3
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.11.0
Jinja2==2.10
lxml==4.1.1
MarkupSafe==1.0
metapub==0.4.3.6
mysqlclient==1.3.12
oauth2client==4.1.2
olefile==0.44
parso==0.1.0
pdfminer2==20151206
pdfx==1.3.0
pexpect==4.3.1
pickleshare==0.7.4
Pillow==4.3.0
pluggy==0.6.0
prompt-toolkit==1.0.15
protobuf==3.5.0.post1
ptyprocess==0.5.2
py==1.5.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
Pygments==2.2.0
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2017.3
requests==2.18.4
requests-file==1.4.2
rsa==3.4.2
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.1.15
swot-python==1.1.1
tabulate==0.8.2
tldextract==2.0.2
tox==2.9.1
traitlets==4.3.2
Unidecode==0.4.21
urllib3==1.22
virtualenv==15.1.0
Wand==0.4.4
wcwidth==0.1.7
Werkzeug==0.13
WTForms==1.0.5
